fix os.remove

+++ email_sender/email_sender.py
@@ -65,7 +65,8 @@ def send_email(recepient, subject, message, qr_file=None):
     mail = smtplib.SMTP_SSL(server)
     mail.login(user, password)
     mail.sendmail(sender, recipient, msg.as_string())
-    os.remove(file_path)
+    if qr_file:
+        os.remove(file_path)
     mail.quit()
